aws-eks/src/functions/ssl-fingerprint/requirements.txt
pyopenssl
certifi
crhelper
boto3
